consolidate directprediction.py
